Herd Immunity Simulator
This is a simulator written in Python demonstrating how Herd Immunity
works.

The user can set various options in the code, including population size,
the amount of people an infected person can go on to infect, the rate of
natural immunity, the rate of immunity among the vaccinated, and what
percentage of the population is vaccinated.

I may expand it in the future to allow users to type in the values as
the program runs, and other features. Feel free to take this and run
with it, including making it more interactive with Windows, Mac, or
Linux.